Refactor key generation and extraction into separate keys module
