+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![feature(error_generic_member_access)]
 
 pub mod errors;
+pub mod keys;
 pub mod protocol;
 pub mod server;
 pub mod storage;
